настраивает получения id пользователя из контекста в api корзины

+++ KartowkaMarkowkaHub.Basket/Controllers/BasketController.cs
@@ -5,12 +5,12 @@ using System.Security.Claims;
 
 namespace KartowkaMarkowkaHub.Basket.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class BasketController : ControllerBase
     {
         private readonly IBasketService _basketService;
-        //private readonly Guid _userId = Guid.Parse("6ebc929b-3785-49d9-9d46-b3b9f70b0bb5");
 
         public BasketController(IBasketService basketService)
         {
@@ -22,7 +22,7 @@ namespace KartowkaMarkowkaHub.Basket.Controllers
         /// </summary>
         /// <param name="userId"></param>
         /// <returns></returns>
-        [Authorize]
+     
         [HttpGet]
         public async Task<IActionResult> Get()
         {            
@@ -38,8 +38,14 @@ namespace KartowkaMarkowkaHub.Basket.Controllers
         /// <param name="productId">id продукта</param>
         /// <returns></returns>
         [HttpPost]
-        public async Task<IActionResult> Post(Guid userId, Guid productId)
+        public async Task<IActionResult> Post(Guid productId)
         {
+            string userIdText = User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? string.Empty;
+
+            if (userIdText == string.Empty)
+                return BadRequest();
+
+            Guid userId = Guid.Parse(userIdText);
             await _basketService.Create(productId, userId);
             return Created();
         }
@@ -50,8 +56,14 @@ namespace KartowkaMarkowkaHub.Basket.Controllers
         /// <param name="productId">id продукта</param>
         /// <returns></returns>
         [HttpDelete]
-        public async Task<IActionResult> Delete(Guid userId, Guid productId)
+        public async Task<IActionResult> Delete(Guid productId)
         {
+            string userIdText = User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? string.Empty;
+
+            if (userIdText == string.Empty)
+                return BadRequest();
+
+            Guid userId = Guid.Parse(userIdText);
             await _basketService.Remove(productId, userId);
             return Ok();
         }
